refactor: move lab-capability cluster to runner-contract (stage 3)
Continues decoupling the optional runner feature from core. Moves the
lab-runner capability cluster (all behavior-free data) into
homeboy-runner-contract:
- RunnerCapabilityPreflight (+ its is_empty() and
  From<PreparedLabRunnerCapability> impls, both contract-local)
- PreparedLabRunnerCapability
- LabRunnerCapabilityContract
- LabRunnerGateMode, LabRunnerGateDecision

The `From<LabRunnerGateDecision> for HomeboyGateResult` impl stays in core
(orphan rule: HomeboyGateResult is core-owned) and now references the
contract type.

All types re-exported from runner::capabilities so internal/CLI call sites
resolve unchanged. External-core consumers (upgrade/runners, rig/capabilities)
repointed to homeboy_runner_contract::, severing those reverse edges. This
also unblocks a future RunnerExecOptions move (its RunnerCapabilityPreflight
field dep is now in the contract; only SourceSnapshot remains).

Part of the staged runner extraction plan (wiki: homeboy-runner-crate-
extraction-plan).

Verified: cargo check clean for -p homeboy-runner-contract, -p homeboy-core
--lib, -p homeboy-cli --lib.
