+++ crates/homeboy-core/src/upgrade/runners/source_checkout.rs
@@ -3,11 +3,11 @@ use super::super::types::InstallMethod;
 use super::*;
 use crate::runner;
 use crate::runner::Runner;
-use crate::runner::RunnerCapabilityPreflight;
 use crate::runner::RunnerExecOptions;
 use crate::runner::RunnerWorkspaceSyncMode;
 use crate::runner::RunnerWorkspaceSyncOptions;
 use crate::Result;
+use homeboy_runner_contract::RunnerCapabilityPreflight;
 use homeboy_runner_contract::RunnerKind;
 use homeboy_runner_contract::RunnerRequiredTool;
 use std::path::Path;
